(1) changed type of dbSize of SDF class to size_t to handle large sdf file (2) added image only mode to 'convert_imageset' tool

+++ SoooA/src/dataset/SDF.h
@@ -44,7 +44,7 @@ private:
 	std::vector<std::string> keys;
 	std::vector<std::string> values;
 
-	int dbSize;
+	size_t dbSize;
 
 	std::ifstream ifs;
 	std::ofstream ofs;
